# Improve ARRAY_FREQUENCY efficiency