Benchmark:

New:
input_bytes:0,output_rows:1,output_bytes:2205
           sql_array_frequencey_10k ::   91.253 cpu ms ::    0B peak memory :: in     1,      0B,      10/s,      0B/s :: out     1,  2.15KB,      10/s,  23.6KB/s
input_bytes:0,output_rows:1,output_bytes:220005
      sql_array_frequencey_all1_10k ::  106.202 cpu ms ::    0B peak memory :: in     1,      0B,       9/s,      0B/s :: out     1,   215KB,       9/s,  1.98MB/s

Old:
input_bytes:0,output_rows:1,output_bytes:2205
           sql_array_frequencey_10k ::  144.473 cpu ms ::    0B peak memory :: in     1,      0B,       6/s,      0B/s :: out     1,  2.15KB,       6/s,  14.9KB/s
input_rows:1,input_bytes:0,output_rows:1,output_bytes:220005
      sql_array_frequencey_all1_10k :: 2006.523 cpu ms ::    0B peak memory :: in     1,      0B,       0/s,      0B/s :: out     1,   215KB,       0/s,   107KB/s
 

Test plan - (Please fill in how you tested your changes)

Please make sure your submission complies with our [Development](https://github.com/prestodb/presto/wiki/Presto-Development-Guidelines#development), [Formatting](https://github.com/prestodb/presto/wiki/Presto-Development-Guidelines#formatting), and [Commit Message](https://github.com/prestodb/presto/wiki/Review-and-Commit-guidelines#commit-formatting-and-pull-requests) guidelines. Don't forget to follow our [attribution guidelines](https://github.com/prestodb/presto/wiki/Review-and-Commit-guidelines#attribution) for any code copied from other projects.

Fill in the release notes towards the bottom of the PR description.
See [Release Notes Guidelines](https://github.com/prestodb/presto/wiki/Release-Notes-Guidelines) for details.


```
== NO RELEASE NOTE ==
```